refactor: update all importation of the ComponentRecommande
